pyproject.toml
[project]
name = "daily-x-signal"
version = "0.1.0"
description = "面向中文用户的 X 高信号日报工具，支持排序、摘要与重点作者发现。"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.32",
]

[project.scripts]
daily-x-signal = "daily_x_signal.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["daily_x_signal"]
